Reorganise spread functions and data
The spread function now uses a single matrix to determine the pixel's vegetation type, instead of one by type.

Fire spread parameters are divided into vegetation intercepts and terrain effects. This will help to ignore some vegetation types when their relative abundance is too low (decreasing the likelihood dimension in some fires).

Intercepts are now parameterised without a reference category to decrease the correlation between them. This is important if we wish to ignore the vegetation type taken as reference, because in that case, the reference intercept would not be identified and parameters would show a terrible correlation. Highly correlated parameters make harder to fit the multivariate normal curve to the simulated likelihood.

Landscapes were re-created to follow the current data structure, so they are no longer compatible with the csv files exported previously for Iván R. to use directly in C++. However, they still take the araucaria dry forests as cypress forests and the steppes as shrublands. The vegetation landscape divide dry forests, but that still has to be done in Google Earth Engine.

TAREAS.txt and notes on estimation.txt were also updated.